First updates to Pharmacies + .gitignore
This is the first commit since sharing the code with aholmes. The only issue might be that he also updated these two files but we can reconcile that. The main goal of this commit is to remove node_modules from github and tracking. Locally, we both need to run these commands:
git rm -r --cached node_modules